requirements.txt
requests
urllib3
pyyaml
